Project code files added
here we have 2 packages service for all services related to student, repository for managing file and data and a class with main method named ApplicationMain . One can just download the code and run ApplicationMain class to run the project. it will ask for username and password, admin and root are the values respectively.